Made compatible with IPython-Notebook 3
The behaviour of install_nbextension changed slightly. In particular

We need to pass in 'user=True' if we want to install to the ~/.ipython directory (by default installation is to /usr/local/share/jupyter  which is not reachable without root access.

We need to pass in javascript files to be install one at a time rather than as a list